# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PTO AI API
  version: 0.1.0
  summary: API сервиса проверки исполнительной документации
  description: |
    Сервис принимает пакеты цифровых документов (PDF с текстовым слоем, DOCX, XLSX/CSV),
    извлекает данные, проверяет их правилами и отдаёт замечания, комплектность и отчёты.

    **Статус: черновик первого этапа.** OCR не входит в этап: документ без текстового слоя
    получает статус `needs_ocr`. Поля, помеченные `x-draft`, требуют согласования между
    фронтендом и бэкендом (см. раздел «Открытые вопросы» ниже).

    ## Как устроен процесс

    1. `POST /objects/{objectId}/packages` создаёт пакет и возвращает ссылки для загрузки файлов.
    2. Клиент загружает каждый файл по `uploadUrl` (PUT).
    3. `POST /packages/{packageId}/start` запускает обработку.
    4. Прогресс приходит по SSE (`GET /packages/{packageId}/events`) или через `GET /packages/{packageId}`.
    5. Результаты читаются из `documents`, `findings`, `checklist`.
    6. Инженер принимает решения по замечаниям (`PATCH /findings/{findingId}`).
    7. `POST /packages/{packageId}/reports` формирует PDF-отчёт для подрядчика.

    ## Этапы обработки

    Каждый документ проходит этапы `extract → ocr → classify → parse → rules → explain`.
    По каждому этапу хранится статус, любой этап можно перезапустить.

    ## Открытые вопросы (x-draft)

    - Формат `bbox` в `SourceRef`: сейчас доли страницы 0..1 от верхнего левого угла.
...
  contact:
    name: PTO AI
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Вход
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: Текущий пользователь
      responses:
        "200":
          description: Пользователь
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /work-types:
    get:
      tags: [Objects]
      operationId: listWorkTypes
      summary: Справочник видов работ
      description: Вид работ определяет шаблон обязательных документов (комплектность).
      responses:
        "200":
          description: Список видов работ
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WorkType" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /objects:
    get:
      tags: [Objects]
      operationId: listObjects
      summary: Список объектов
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: q
          in: query
          description: Поиск по названию, адресу, подрядчику
          schema: { type: string }
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, -name, readiness, -readiness, lastCheckedAt, -lastCheckedAt]
            default: -lastCheckedAt
      responses:
        "200":
          description: Страница объектов
          content:
            application/json:
              schema:
                type: object
                required: [items, meta]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ConstructionObject" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Objects]
      operationId: createObject
      summary: Создать объект
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ObjectInput" }
      responses:
        "201":
          description: Объект создан
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConstructionObject" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /objects/{objectId}:
    parameters:
      - $ref: "#/components/parameters/ObjectId"
    get:
      tags: [Objects]
      operationId: getObject
      summary: Объект
      responses:
        "200":
          description: Объект
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConstructionObject" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Objects]
      operationId: updateObject
      summary: Изменить объект
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ObjectUpdate" }
      responses:
        "200":
          description: Объект обновлён
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConstructionObject" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      tags: [Objects]
      operationId: deleteObject
      summary: Удалить объект
      responses:
        "204": { description: Объект удалён }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /objects/{objectId}/packages:
    parameters:
      - $ref: "#/components/parameters/ObjectId"
    get:
      tags: [Packages]
      operationId: listPackages
      summary: Пакеты объекта
      description: Пакет это версия загрузки. Список отсортирован от новых к старым.
      responses:
        "200":
          description: Пакеты
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Package" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Packages]
      operationId: createPackage
      summary: Создать пакет и получить ссылки для загрузки
      description: |
        Принимает описание файлов, создаёт `Package` и `Document` для каждого файла
        и возвращает временные ссылки для загрузки в хранилище.
        Дубликаты по `sha256` помечаются в ответе (`duplicateOf`), повторно не загружаются.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PackageCreate" }
      responses:
        "201":
          description: Пакет создан
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PackageCreated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413":
          description: Файл или пакет превышает допустимый размер
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "415":
          description: Неподдерживаемый формат файла
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /packages/{packageId}:
    parameters:
      - $ref: "#/components/parameters/PackageId"
    get:
      tags: [Packages]
      operationId: getPackage
      summary: Состояние пакета
      description: Статус, общий прогресс и документы с этапами. Подходит для поллинга, если SSE недоступен.
      responses:
        "200":
          description: Пакет
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PackageDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
  /packages/{packageId}/start:
    parameters:
      - $ref: "#/components/parameters/PackageId"
    post:
      tags: [Packages]
      operationId: startPackage
      summary: Запустить обработку
      description: Вызывается после того, как все файлы загружены. Идемпотентен.
      responses:
        "202":
          description: Обработка поставлена в очередь
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Package" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Не все файлы загружены
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /packages/{packageId}/events:
    parameters:
      - $ref: "#/components/parameters/PackageId"
    get:
      tags: [Packages]
      operationId: streamPackageEvents
      summary: Поток событий обработки (SSE)
      x-draft: true
      description: |
        Server-Sent Events. Соединение остаётся открытым до завершения пакета.
        Каждое событие содержит JSON по схеме `PackageEvent`. Имя события (`event:`) совпадает
        с полем `type`. При подключении сервер сначала отправляет `package.snapshot`
        с текущим состоянием. Так как `EventSource` не умеет передавать заголовки,
        для авторизации допустим короткоживущий токен в `access_token` или cookie-сессия.
      parameters:
        - name: access_token
          in: query
          required: false
          description: Токен для клиентов, которые не могут задать заголовок Authorization
          schema: { type: string }
      responses:
        "200":
          description: Поток событий
          content:
            text/event-stream:
              schema: { $ref: "#/components/schemas/PackageEvent" }
        "404": { $ref: "#/components/responses/NotFound" }
  /packages/{packageId}/reports:
    parameters:
      - $ref: "#/components/parameters/PackageId"
    post:
      tags: [Reports]
      operationId: createReport
      summary: Сформировать отчёт по пакету
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReportCreate" }
      responses:
        "202":
          description: Генерация запущена
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Обработка пакета ещё не завершена
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /objects/{objectId}/reports:
    parameters:
      - $ref: "#/components/parameters/ObjectId"
    get:
      tags: [Reports]
      operationId: listReports
      summary: Отчёты объекта
      responses:
        "200":
          description: Отчёты
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Report" }
        "404": { $ref: "#/components/responses/NotFound" }
  /reports/{reportId}:
    parameters:
      - $ref: "#/components/parameters/ReportId"
    get:
      tags: [Reports]
      operationId: getReport
      summary: Отчёт
      responses:
        "200":
          description: Отчёт
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "404": { $ref: "#/components/responses/NotFound" }
  /reports/{reportId}/download:
    parameters:
      - $ref: "#/components/parameters/ReportId"
    get:
      tags: [Reports]
      operationId: downloadReport
      summary: Скачать PDF отчёта
      responses:
        "200":
          description: PDF-файл
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Отчёт ещё не готов
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-draft: true
  parameters:
    ObjectId:
      name: objectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PackageId:
      name: packageId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ReportId:
      name: reportId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { code: unauthorized, message: Требуется вход в систему }
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { code: not_found, message: Ресурс не найден }
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            code: validation_error
            message: Проверьте введённые данные
            details:
              - { field: name, message: Обязательное поле }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Машинный код ошибки
          examples: [validation_error]
        message:
          type: string
          description: Понятное человеку описание, что случилось и что делать
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    PageMeta:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
    SeverityCounts:
      type: object
      required: [error, warning, info, critical]
      properties:
        error: { type: integer, minimum: 0 }
        warning: { type: integer, minimum: 0 }
        info: { type: integer, minimum: 0 }
        critical: { type: integer, minimum: 0 }

    # ---------- Перечисления ----------
...
    DocumentType:
      type: string
      x-draft: true
      enum:
        - aosr
        - general_work_log
        - concrete_log
        - welding_log
        - material_passport
        - as_built_scheme
        - test_report
        - other
        - unknown
      description: |
        aosr: акт освидетельствования скрытых работ; general_work_log: общий журнал работ;
        concrete_log: журнал бетонных работ; welding_log: журнал сварки;
        material_passport: паспорт материалов; as_built_scheme: исполнительная схема;
        test_report: протокол испытаний; unknown: тип не определён.
    DocumentStatus:
      type: string
      enum: [awaiting_upload, uploaded, queued, processing, done, failed, needs_ocr]
      description: |
        Статусы хранения/очереди совпадают с Prisma DocumentStatus.
        needs_ocr: API-флаг для файла без текстового слоя (скан); на первом этапе не обрабатывается.
    Stage:
      type: string
      enum: [extract, ocr, classify, parse, rules, explain]
      description: Совпадает с Prisma StageName (этап report в БД отсутствует)
    StageStatus:
      type: string
      enum: [pending, running, succeeded, failed, skipped]
      description: Совпадает с Prisma StageStatus (succeeded вместо done)
    PackageStatus:
      type: string
      enum: [uploading, queued, processing, done, partial, failed]
      description: partial означает, что часть документов не удалось обработать, остальные готовы
...
    ReportStatus:
      type: string
      enum: [queued, generating, ready, failed]
      description: Совпадает с Prisma ReportStatus

    # ---------- Пользователь ----------
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    AuthResponse:
      type: object
      required: [accessToken, expiresIn, user]
      properties:
        accessToken: { type: string }
        expiresIn: { type: integer, description: Время жизни токена, секунды }
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        role:
          type: string
          enum: [engineer, contractor]

    # ---------- Объекты ----------
    WorkType:
      type: object
      required: [id, name]
      properties:
        id: { type: string, examples: [monolith] }
        name: { type: string, examples: [Монолитные работы] }
    ObjectInput:
      type: object
      required: [name, workTypeId]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        address: { type: string }
        customer: { type: string, description: Заказчик }
        contractor: { type: string, description: Подрядчик }
        workTypeId: { type: string }
    ObjectUpdate:
      type: object
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        address: { type: string }
        customer: { type: string }
        contractor: { type: string }
        workTypeId: { type: string }
    ConstructionObject:
      allOf:
        - $ref: "#/components/schemas/ObjectInput"
        - type: object
          required: [id, readiness, findings, createdAt]
          properties:
            id: { type: string, format: uuid }
            readiness:
              type: number
              minimum: 0
              maximum: 1
              description: Готовность пакета документов, доля от 0 до 1
            findings: { $ref: "#/components/schemas/SeverityCounts" }
            lastCheckedAt: { type: [string, "null"], format: date-time }
            createdAt: { type: string, format: date-time }

    # ---------- Пакеты ----------
    PackageCreate:
      type: object
      required: [files]
      properties:
        files:
          type: array
          minItems: 1
          maxItems: 200
          items:
            type: object
            required: [fileName, mimeType, sizeBytes]
            properties:
              fileName: { type: string }
              mimeType:
                type: string
                enum:
                  - application/pdf
                  - application/vnd.openxmlformats-officedocument.wordprocessingml.document
                  - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
                  - text/csv
              sizeBytes: { type: integer, minimum: 1 }
              sha256:
                type: string
                pattern: "^[a-f0-9]{64}$"
                description: Хэш файла для дедупликации (необязательно)
    PackageCreated:
      type: object
      required: [package, uploads]
      properties:
        package: { $ref: "#/components/schemas/Package" }
        uploads:
          type: array
          items:
            type: object
            required: [documentId, fileName]
            properties:
              documentId: { type: string, format: uuid }
              fileName: { type: string }
              uploadUrl:
                type: [string, "null"]
                format: uri
                description: Ссылка для PUT-загрузки. Пусто, если файл дубликат.
              headers:
                type: object
                additionalProperties: { type: string }
                description: Заголовки, которые нужно передать при загрузке
              expiresAt: { type: string, format: date-time }
              duplicateOf:
                type: [string, "null"]
                format: uuid
                description: Идентификатор уже загруженного документа с тем же хэшем
    Package:
      type: object
      required: [id, objectId, version, status, progress, createdAt]
      properties:
        id: { type: string, format: uuid }
        objectId: { type: string, format: uuid }
        version:
          type: integer
          minimum: 1
          description: Порядковый номер проверки на объекте
        status: { $ref: "#/components/schemas/PackageStatus" }
        progress:
          type: number
          minimum: 0
          maximum: 1
        documentsCount: { type: integer, minimum: 0 }
        findings: { $ref: "#/components/schemas/SeverityCounts" }
        createdAt: { type: string, format: date-time }
        finishedAt: { type: [string, "null"], format: date-time }
    PackageDetail:
      allOf:
        - $ref: "#/components/schemas/Package"
        - type: object
          required: [documents]
          properties:
            documents:
              type: array
              items: { $ref: "#/components/schemas/Document" }
    PackageEvent:
      description: |
        Событие SSE. Поле `type` совпадает с именем события.
        - `package.snapshot`: текущее состояние при подключении
        - `document.stage`: изменился статус этапа документа
        - `package.progress`: изменился общий прогресс
        - `package.done`: обработка завершена (статус `done`, `partial` или `failed`)
      oneOf:
        - $ref: "#/components/schemas/PackageSnapshotEvent"
        - $ref: "#/components/schemas/DocumentStageEvent"
        - $ref: "#/components/schemas/PackageProgressEvent"
        - $ref: "#/components/schemas/PackageDoneEvent"
      discriminator:
        propertyName: type
        mapping:
          package.snapshot: "#/components/schemas/PackageSnapshotEvent"
          document.stage: "#/components/schemas/DocumentStageEvent"
          package.progress: "#/components/schemas/PackageProgressEvent"
          package.done: "#/components/schemas/PackageDoneEvent"
    PackageSnapshotEvent:
      type: object
      required: [type, package]
      properties:
        type: { type: string, const: package.snapshot }
        package: { $ref: "#/components/schemas/PackageDetail" }
    DocumentStageEvent:
      type: object
      required: [type, documentId, stage, status]
      properties:
        type: { type: string, const: document.stage }
        documentId: { type: string, format: uuid }
        stage: { $ref: "#/components/schemas/Stage" }
        status: { $ref: "#/components/schemas/StageStatus" }
        error: { $ref: "#/components/schemas/StageError" }
    PackageProgressEvent:
      type: object
      required: [type, progress]
      properties:
        type: { type: string, const: package.progress }
        progress: { type: number, minimum: 0, maximum: 1 }
    PackageDoneEvent:
      type: object
      required: [type, status]
      properties:
        type: { type: string, const: package.done }
        status:
          type: string
          enum: [done, partial, failed]
        findings: { $ref: "#/components/schemas/SeverityCounts" }

    # ---------- Документы ----------
    StageError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [no_text_layer, corrupted_file, llm_unavailable]
        message:
          type: string
          description: Причина простыми словами, показывается пользователю
    StageState:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/StageStatus" }
        error: { $ref: "#/components/schemas/StageError" }
        finishedAt: { type: [string, "null"], format: date-time }
    Document:
      type: object
      required: [id, objectId, packageId, fileName, mimeType, sizeBytes, type, status, stages, findings, createdAt]
      properties:
        id: { type: string, format: uuid }
        objectId: { type: string, format: uuid }
        packageId: { type: string, format: uuid }
        fileName: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        pageCount:
          type: [integer, "null"]
          description: Для PDF; для XLSX число листов
        type: { $ref: "#/components/schemas/DocumentType" }
        typeConfidence:
          type: [number, "null"]
          minimum: 0
          maximum: 1
        typeEditedManually: { type: boolean, default: false }
        status: { $ref: "#/components/schemas/DocumentStatus" }
        stages:
          type: object
          description: Состояние каждого этапа обработки
          properties:
            extract: { $ref: "#/components/schemas/StageState" }
            ocr: { $ref: "#/components/schemas/StageState" }
            classify: { $ref: "#/components/schemas/StageState" }
            parse: { $ref: "#/components/schemas/StageState" }
            rules: { $ref: "#/components/schemas/StageState" }
            explain: { $ref: "#/components/schemas/StageState" }
        findings: { $ref: "#/components/schemas/SeverityCounts" }
        createdAt: { type: string, format: date-time }
...
    ReportCreate:
      type: object
      properties:
        scope:
          type: string
          enum: [accepted, errors, all_open]
          default: accepted
          description: Что включить, принятые замечания, только ошибки или все открытые
        recipient:
          type: string
          description: Кому адресован отчёт (подрядчик)
    Report:
      type: object
      required: [id, objectId, packageId, status, version, createdAt]
      properties:
        id: { type: string, format: uuid }
        objectId: { type: string, format: uuid }
        packageId: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/ReportStatus" }
        version: { type: integer, minimum: 1 }
        scope: { type: string, enum: [accepted, errors, all_open] }
        recipient: { type: [string, "null"] }
        findingsCount: { type: integer, minimum: 0 }
        error: { $ref: "#/components/schemas/StageError" }
        createdBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        readyAt: { type: [string, "null"], format: date-time }